Add the initial verison of my NES helper scripts for Ghidra.
This introduces two scripts:

* export_nes.py
* populate_ppu_tile_16.py

The main script, `export_nes.py`, is built to turn a disassembled NES
file into both a compilable plain-text NES assembly file and a browsable
HTML file.

`populate_ppu_tile_16.py` takes a 16-byte range of PPU tile data and
turns it into an ASCII visualization of the tile that can be embedded
in a comment.