Add data value references to end-of-line comments.
When a data value has a reference set to another label, whether the
address at the label is a direct memory match of inferred through
another process, the referenced label will now be shown in an
end-of-line comment and linked in the HTML output.

This can help with documenting tables pointing to other state using
relative offsets or those that are otherwise transformed.